- split application into blueprints - create users folder in root folder - create posts folder in root folder - create main folder in root folder - create init files in posts, main and users folders - create routes.py file in posts, main and users folders - create forms.py file in users and posts folders - create config file for configurations - move SECRET_KEY and SQLALCHEMY_DATABASE_URI to environment variables
