+++ nba_homegames/__init__.py
@@ -3,23 +3,21 @@ from flask_sqlalchemy import SQLAlchemy
 from flask_bcrypt import Bcrypt
 from flask_login import LoginManager
 from flask_mail import Mail
-import os
+from nba_homegames.config import Config
 
 app = Flask(__name__)
-app.config['SECRET_KEY'] = 'fac97d332b5b472f5f509009eba1b419'
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///site.db'  #  Relative path from this file
+app.config.from_object(Config)
 db = SQLAlchemy(app)  #  SQLAlchemy database instance
 bcrypt = Bcrypt(app)
 login_manager = LoginManager(app)
-login_manager.login_view = 'login'
+login_manager.login_view = 'users.login'
 login_manager.login_message_category = 'info'
-#  Constants for sending email
-app.config['MAIL_SERVER'] = 'smtp.gmail.com'
-app.config['MAIL_PORT'] = 587
-app.config['MAIL_USE_TLS'] = True
-#  Environment variables for sensitive information
-app.config['MAIL_USERNAME'] = os.environ.get('EMAIL_USER')
-app.config['MAIL_PASSWORD'] = os.environ.get('EMAIL_PASS')
 mail = Mail(app)
 
-from nba_homegames import routes
+from nba_homegames.users.routes import users
+from nba_homegames.posts.routes import posts
+from nba_homegames.main.routes import main
+
+app.register_blueprint(users)
+app.register_blueprint(posts)
+app.register_blueprint(main)
