+++ nba_homegames/templates/account.html
@@ -59,7 +59,7 @@
     </form>
     <div class="border-top pt-3">
         <small class="text-muted">
-            Already have an account? <a class="ml-2" href="{{url_for('login')}}"> Sign in</a>
+            Already have an account? <a class="ml-2" href="{{url_for('users.login')}}"> Sign in</a>
         </small>
     </div>
 </div>
