+++ nba_homegames/templates/login.html
@@ -42,14 +42,14 @@
         <div class="form-group">
             {{form.submit(class = "btn btn-outline-info")}}
             <small class = "text-muted ml-2">
-                <a href = "{{url_for('reset_request')}}"> Forgot Password?</a>
+                <a href = "{{url_for('users.reset_request')}}"> Forgot Password?</a>
             </small>
         </div>
     </form>
 </div>
 <div class = "border-top pt-3">
     <small class = "text-muted">
-        Need an account? <a class = "ml-2" href = "{{url_for('register')}}"> Sign up here</a>
+        Need an account? <a class = "ml-2" href = "{{url_for('users.register')}}"> Sign up here</a>
     </small>
 </div>
 {% endblock content %}
