+++ nba_homegames/templates/post.html
@@ -5,12 +5,12 @@
         <img class = "rounded-circle article-img" src = "{{ url_for('static', filename = 'profile_pics/' + post.author.image_file) }}">
         <div class="media-body">
             <div class="article-metadata">
-                <a class="mr-2" href="{{ url_for('user_posts', username = post.author.username) }}">{{ post.author.username }}</a>
+                <a class="mr-2" href="{{ url_for('users.user_posts', username = post.author.username) }}">{{ post.author.username }}</a>
                 <small class="text-muted">{{ post.date_posted.strftime('%Y-%m-%d') }}</small>
                 <!-- Only let update and delete be viewable by logged in user -->
                 {% if post.author == current_user %}
                     <div>
-                        <a class = "btn btn-secondary btn-sm mt-1 mb-1" href = "{{url_for('update_post', post_id = post.id)}}">Update</a>
+                        <a class = "btn btn-secondary btn-sm mt-1 mb-1" href = "{{url_for('posts.update_post', post_id = post.id)}}">Update</a>
                         <button type = "button" class = "btn btn-danger btn-sm m-1" data-toggle="modal" data-target="#deleteModal">Delete</button>
                     </div>
                 {% endif %}
@@ -32,7 +32,7 @@
                 </div>
                 <div class="modal-footer">
                     <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-                    <form action = "{{ url_for('delete_post', post_id = post.id) }}" method = "POST">
+                    <form action = "{{ url_for('posts.delete_post', post_id = post.id) }}" method = "POST">
                         <input class = "btn btn-danger" type = "submit" value = "Delete">
                     </form>
                 </div>
