+++ nba_homegames/templates/register.html
@@ -72,7 +72,7 @@
 </div>
 <div class = "border-top pt-3">
     <small class = "text-muted">
-        Already have an account? <a class = "ml-2" href = "{{url_for('login')}}"> Sign in</a>
+        Already have an account? <a class = "ml-2" href = "{{url_for('users.login')}}"> Sign in</a>
     </small>
 </div>
 {% endblock content %}
