+++ nba_homegames/templates/user_posts.html
@@ -7,10 +7,10 @@
             <img class = "rounded-circle article-img" src = "{{ url_for('static', filename = 'profile_pics/' + post.author.image_file) }}">
             <div class="media-body">
                 <div class="article-metadata">
-                    <a class="mr-2" href="{{ url_for('user_posts', username = post.author.username) }}">{{ post.author.username }}</a>
+                    <a class="mr-2" href="{{ url_for('users.user_posts', username = post.author.username) }}">{{ post.author.username }}</a>
                     <small class="text-muted">{{ post.date_posted.strftime('%Y-%m-%d') }}</small>
                 </div>
-                <h2><a class="article-title" href="{{ url_for('post', post_id = post.id) }}">{{ post.title }}</a></h2>
+                <h2><a class="article-title" href="{{ url_for('posts.post', post_id = post.id) }}">{{ post.title }}</a></h2>
                 <p class="article-content">{{ post.content }}</p>
             </div>
         </article>
@@ -19,9 +19,9 @@
         {% if page_num %}
             {% if posts.page == page_num %}
                 <!-- Make current page number at bottom filled in -->
-                <a class = "btn btn-info mb-4" href = "{{url_for('user_posts', username = user.username, page = page_num)}}">{{ page_num }}</a>
+                <a class = "btn btn-info mb-4" href = "{{url_for('users.user_posts', username = user.username, page = page_num)}}">{{ page_num }}</a>
             {% else %}
-                <a class = "btn btn-outline-info mb-4" href = "{{url_for('home', username = user.username, page = page_num)}}">{{ page_num }}</a>
+                <a class = "btn btn-outline-info mb-4" href = "{{url_for('main.home', username = user.username, page = page_num)}}">{{ page_num }}</a>
             {% endif %}
         {% else %}
             ...
