+++ nba_homegames/users/forms.py
@@ -1,9 +1,10 @@
+
 from flask_wtf import FlaskForm
 from flask_wtf.file import FileField, FileAllowed
-from wtforms import StringField, PasswordField, SubmitField, BooleanField, TextAreaField
+from wtforms import StringField, PasswordField, SubmitField, BooleanField
 from wtforms.validators import DataRequired, Length, Email, EqualTo, ValidationError
-from nba_homegames.models import User
 from flask_login import current_user
+from nba_homegames.models import User
 
 # Class for Registration Form, inherits from FlaskForm
 class RegistrationForm(FlaskForm):
@@ -64,11 +65,6 @@ class UpdateAccountForm(FlaskForm):
             #  If email already exists, throw validation error
             if user:
                 raise ValidationError('Email entered is already taken. Please choose another one.')
-
-class PostForm(FlaskForm):
-    title = StringField('Title', validators = [DataRequired()])
-    content = TextAreaField('Content', validators = [DataRequired()])
-    submit = SubmitField('Post')
 
 class RequestResetForm(FlaskForm):
     email = StringField('Email',
